Extract common logic
