+++ app/src/main/java/com/nextcloud/client/jobs/notification/WorkerNotificationManager.kt
@@ -7,6 +7,7 @@
 
 package com.nextcloud.client.jobs.notification
 
+import android.app.Notification
 import android.app.NotificationManager
 import android.content.Context
 import android.graphics.BitmapFactory
@@ -17,22 +18,38 @@ import com.owncloud.android.ui.notifications.NotificationUtils
 import com.owncloud.android.utils.theme.ViewThemeUtils
 
 open class WorkerNotificationManager(
+    private val id: Int,
     private val context: Context,
-    private val viewThemeUtils: ViewThemeUtils
+    viewThemeUtils: ViewThemeUtils,
+    private val tickerId: Int
 ) {
     var currentOperationTitle: String? = null
 
     val notificationManager = context.getSystemService(Context.NOTIFICATION_SERVICE) as NotificationManager
 
-    fun getNotificationBuilder(tickerId: Int): NotificationCompat.Builder {
-        return NotificationUtils.newNotificationBuilder(context, viewThemeUtils).apply {
-            setTicker(context.getString(tickerId))
-            setSmallIcon(R.drawable.notification_icon)
-            setLargeIcon(BitmapFactory.decodeResource(context.resources, R.drawable.notification_icon))
+    var notificationBuilder: NotificationCompat.Builder = NotificationUtils.newNotificationBuilder(context, viewThemeUtils).apply {
+        setTicker(context.getString(tickerId))
+        setSmallIcon(R.drawable.notification_icon)
+        setLargeIcon(BitmapFactory.decodeResource(context.resources, R.drawable.notification_icon))
 
-            if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.O) {
-                setChannelId(NotificationUtils.NOTIFICATION_CHANNEL_DOWNLOAD)
-            }
+        if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.O) {
+            setChannelId(NotificationUtils.NOTIFICATION_CHANNEL_DOWNLOAD)
         }
+    }
+
+    fun showNotification() {
+        notificationManager.notify(id, notificationBuilder.build())
+    }
+
+    fun dismissWorkerNotifications() {
+        notificationManager.cancel(id)
+    }
+
+    fun getId(): Int {
+        return id
+    }
+
+    fun getNotification(): Notification {
+        return notificationBuilder.build()
     }
 }
